LayoutTests/http/tests/security/contentSecurityPolicy/script-loads-with-img-src-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'script-img'.



--------
Frame: '<!--frame1-->'
--------
PASS
